rename catalog_types.py to catalog_utils.py, trim docstring
